scammer-profiler/requirements.txt
pandas
numpy
scikit-learn
umap-learn
hdbscan
spacy
langdetect
nltk
textblob
networkx
python-louvain
streamlit
plotly
pyvis
